Added exception casees to the api

+++ api/views.py
@@ -1,8 +1,10 @@
+from urllib import request, response
 from django.shortcuts import render
 from django.http import JsonResponse
 
 from rest_framework.decorators import api_view
 from rest_framework.response import Response
+from rest_framework import status
 from .serializers import TaskSerializer
 
 from .models import Task
@@ -23,16 +25,27 @@ def apiOverview(request):
 
 @api_view(['GET'])
 def taskList(request):
-	tasks = Task.objects.all()
-	serializer = TaskSerializer(tasks, many=True)
-	return Response(serializer.data)
+	try:
+		tasks = Task.objects.all()
+	except Task.DoesNotExist:
+		return Response(status=status.HTTP_404_NOT_FOUND)
+
+	if request.method == "GET":
+		serializer = TaskSerializer(tasks, many=True)
+		return Response(serializer.data)
 
 
 @api_view(['GET'])
 def taskDetail(request, pk):
-	tasks = Task.objects.get(id=pk)
-	serializer = TaskSerializer(tasks, many=False)
-	return Response(serializer.data)
+
+	try:
+		tasks = Task.objects.get(id=pk)
+	except Task.DoesNotExist:
+		return Response(status=status.HTTP_404_NOT_FOUND)
+	
+	if request.method == "GET":
+		serializer = TaskSerializer(tasks, many=False)
+		return Response(serializer.data)
 
 
 @api_view(['POST'])
@@ -45,21 +58,37 @@ def taskCreate(request):
     return Response(serializer.data)
 
 
-@api_view(['GET','POST'])
+@api_view(['PUT'])
 def taskUpdate(request, pk):
-    task = Task.objects.get(id=pk)
-    serializer = TaskSerializer(instance=task, data=request.data)
+
+	try:
+		task = Task.objects.get(id=pk)
+	except task.DoesNotExist:
+		return response(status=status.HTTP_404_NOT_FOUND)
     
-    if serializer.is_valid():
-        serializer.save()
-    
-    return Response(serializer.data)
+	if request.method == "PUT":
+		serializer = TaskSerializer(task, data=request.data)
+		data={}
+		if serializer.is_valid():
+			serializer.save()
+			data["success"] = "update succesful"
+			return Response(data=data)
+		return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
 
-@api_view(['GET','DELETE'])
+@api_view(['DELETE'])
 def taskDelete(request, pk):
-    task = Task.objects.get(id=pk)
+	
+	try:
+		task = Task.objects.get(id=pk)
+	except Task.DoesNotExist:
+		return Response(status=status.HTTP_404_NOT_FOUND)
     
-    task.delete()
-    
-    return Response('Item succsesfully delete!')
+	if request.method == "DELETE":
+		operation = task.delete()
+		data={}
+		if operation:
+			data["success"] = "delete succesful"
+		else:
+			data["failure"] = "delete failed"
+		return Response(data=data)
